refactor code for modal
