+++ src/components/Modal/BaseModal.js
@@ -1,50 +1,20 @@
 import React from 'react';
-import PropTypes from 'prop-types';
 import {View} from 'react-native';
 import ReactNativeModal from 'react-native-modal';
 import {SafeAreaInsetsContext} from 'react-native-safe-area-context';
-import CustomStatusBar from './CustomStatusBar';
-import KeyboardShortcut from '../libs/KeyboardShortcut';
-import styles, {getSafeAreaPadding} from '../styles/styles';
-import themeColors from '../styles/themes/default';
-import getModalStyles from '../styles/getModalStyles';
-import CONST from '../CONST';
-import withWindowDimensions, {windowDimensionsPropTypes} from './withWindowDimensions';
+import CustomStatusBar from '../CustomStatusBar';
+import KeyboardShortcut from '../../libs/KeyboardShortcut';
+import styles, {getSafeAreaPadding} from '../../styles/styles';
+import themeColors from '../../styles/themes/default';
+import {baseModalPropTypes} from './ModalPropTypes';
 
-const propTypes = {
-    // Callback method fired when the user requests to close the modal
-    onClose: PropTypes.func.isRequired,
-
-    // State that determines whether to display the modal or not
-    isVisible: PropTypes.bool.isRequired,
-
-    // Modal contents
-    children: PropTypes.node.isRequired,
-
-    // Callback method fired when the user requests to submit the modal content.
-    onSubmit: PropTypes.func,
-
-    // Callback method fired when the modal is hidden
-    onModalHide: PropTypes.func,
-
-    // Style of modal to display
-    type: PropTypes.oneOf([
-        CONST.MODAL.MODAL_TYPE.CENTERED,
-        CONST.MODAL.MODAL_TYPE.BOTTOM_DOCKED,
-        CONST.MODAL.MODAL_TYPE.POPOVER,
-        CONST.MODAL.MODAL_TYPE.RIGHT_DOCKED,
-    ]),
-
-    ...windowDimensionsPropTypes,
-};
 
 const defaultProps = {
     onSubmit: null,
-    type: '',
     onModalHide: () => {},
 };
 
-const Modal = (props) => {
+const BaseModal = (props) => {
     const subscribeToKeyEvents = () => {
         KeyboardShortcut.subscribe('Escape', props.onClose, [], true);
         KeyboardShortcut.subscribe('Enter', props.onSubmit, [], true);
@@ -63,11 +33,7 @@ const Modal = (props) => {
         shouldAddTopSafeAreaPadding,
         shouldAddBottomSafeAreaPadding,
         hideBackdrop,
-    } = getModalStyles(props.type, {
-        windowWidth: props.windowWidth,
-        windowHeight: props.windowHeight,
-        isSmallScreenWidth: props.isSmallScreenWidth,
-    });
+    } = props;
     return (
         <ReactNativeModal
             onBackdropPress={props.onClose}
@@ -120,7 +86,7 @@ const Modal = (props) => {
     );
 };
 
-Modal.propTypes = propTypes;
-Modal.defaultProps = defaultProps;
-Modal.displayName = 'Modal';
-export default withWindowDimensions(Modal);
+BaseModal.propTypes = baseModalPropTypes;
+BaseModal.defaultProps = defaultProps;
+BaseModal.displayName = 'BaseModal';
+export default BaseModal;
